requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.4
transformers>=4.30.2
accelerate>=0.21.0
safetensors>=0.3.1
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.0
scipy>=1.10.0
huggingface-hub>=0.16.4
controlnet_aux>=0.0.6
mediapipe>=0.10.3
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6 